requirements.txt
streamlit
langchain
langchain-community
langchain-core
mysql-connector-python
pymysql
langchain-google-genai
python-dotenv
ipykernel
fastapi 
uvicorn 
pydantic
gradio